requirements.txt
starlette==0.52.1
uvicorn[standard]==0.41.0
deepgram-sdk==6.0.0
twilio==9.10.2
python-dotenv==1.2.1
python-multipart==0.0.22
